Func. getDatasets moved to procs/adapter.py
